Gas optimizations (#37)
* Emission Rate is uint64

* Time is uint64, Epoch is uint8

+++ contracts/EmissionRates.sol
@@ -16,7 +16,7 @@ pragma solidity ^0.8;
  * @notice Contains precalculated emission rated for 120 epochs to prevent complex floating point calculations
  */
 abstract contract EmissionRates {
-    uint256[120] internal _rates = [
+    uint64[120] internal _rates = [
         10000000000000000000,
         9682601515616320000,
         9375277211021540000,
@@ -139,7 +139,7 @@ abstract contract EmissionRates {
         215305447128642020
     ];
 
-    function rates(uint256 _epoch) external view returns (uint256) {
+    function rates(uint8 _epoch) external view returns (uint256) {
         if (_epoch < _rates.length) {
             return _rates[_epoch];
         } else {
